requirements.txt
asgiref==3.5.0
attrs==22.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cssutils==2.6.0
currency-symbols==2.0.3
cycler==0.11.0
distlib==0.3.4
Django==4.0.3
django-chart-js==0.0.3
django-chartjs==2.3.0
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.5
EasyProcess==1.1
entrypoint2==1.0
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.6.0
Flask==2.2.3
fonttools==4.38.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
lasio==0.30
MarkupSafe==2.1.2
matplotlib==3.7.0
multipledispatch==1.0.0
numpy==1.24.1
openpyxl==3.0.10
packaging==23.0
Pillow==9.4.0
platformdirs==2.5.1
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.1
python-dateutil==2.8.2
pytz==2022.7.1
pyunpack==0.2.2
pywin32==303
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
schedule==1.2.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.66.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.14
virtualenv==20.14.1
waitress==2.1.2
webcolors==1.12
Werkzeug==2.2.3
xlrd==2.0.1
xls2xlsx==0.2.0
